requirements.txt
aniso8601==9.0.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.1.2
blinker==1.6.2
cachelib==0.9.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.36.0
colorama==0.4.6
configparser==5.3.0
contextlib2==21.6.0
coverage==7.2.7
crypto==1.4.1
cryptography==41.0.1
DateTime==5.1
Deprecated==1.2.14
enum34==1.1.10
extras==1.0.0
fixtures==4.1.0
Flask==2.2.3
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Limiter==3.3.1
Flask-SQLAlchemy==3.0.3
funcsigs==1.0.2
geographiclib==2.0
geopy==2.4.1
greenlet==2.0.2
idna==2.7
importlib-metadata==6.4.1
importlib-resources==5.12.0
ipaddress==1.0.23
itsdangerous==2.0.1
Jinja2==3.1.2
jsonschema==4.17.3
limits==3.5.0
linecache2==1.0.0
logger==1.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
Naked==0.1.32
nose==1.3.7
numpy==1.24.4
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
paramiko==3.2.0
pbr==5.11.1
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
Pygments==2.15.1
PyJWT==2.6.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pypi==2.1
pyrsistent==0.19.3
pyshorteners==1.0.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-mimeparse==1.6.0
python-subunit==1.4.2
python-toolbox==1.0.11
pytz==2023.3
PyYAML==6.0
qrcode==6.0
redis==4.5.5
reportlab==4.0.8
requests==2.28.2
rich==13.4.1
rsa==4.9
schema==0.7.5
shellescape==3.8.1
six==1.16.0
SQLAlchemy==2.0.9
testtools==2.6.0
traceback2==1.4.0
typing==3.7.4.3
typing_extensions==4.5.0
tzdata==2023.4
urllib3==1.26.5
Werkzeug==2.3.4
wrapt==1.15.0
x509==0.1
zipp==3.15.0
zope.interface==6.0